+++ docs/release-notes.md
@@ -21,6 +21,7 @@
 * For modders:
   * Added asset propagation for `Data\Concessions`.
   * Added SMAPI version and bitness to the console title before startup to simplify troubleshooting.
+  * If a map loads a tilesheet path with no file extension, SMAPI now automatically links it to a `.png` version in the map folder if possible.
   * Improved error-handling during asset propagation.
   * Fixed `Context.IsMainPlayer` returning true for a farmhand in split-screen mode before the screen is initialized.
   * Fixed error when editing bundle data while a split-screen player is joining.